Update sketch.js
Fixed a bug with the double-click function: after adding new lines, repeatedly double-clicking the same area no longer causes the rectangle to be split into two new ones by the underlying line. Also adjusted the code so that once a rectangle is turned white, it can no longer be recolored.